# laravelapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for your Laravel backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/coupon": {
            "get": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Get all coupons",
                "operationId": "b2959e07dc9e4c1b5a95f17483828a58",
                "responses": {
                    "200": {
                        "description": "List of coupons"
                    }
                }
            },
            "post": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Create a new coupon",
                "operationId": "bbbced0aac29488122138213afcfae93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "expires_at": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Coupon created"
                    }
                }
            }
        },
        "/api/coupon/{id}": {
            "get": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Get a coupon by ID",
                "operationId": "d7eae06c67fadc363b38a418c0330986",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupon found"
                    },
                    "404": {
                        "description": "Coupon not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Delete a coupon",
                "operationId": "777a8e66a00b38ecd125aa490b5ae4ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupon delete successful"
                    },
                    "404": {
                        "description": "Coupon not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Update an existing coupon",
                "operationId": "dd1e58b1967784966ae83657db699498",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "expires_at": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coupon updated"
                    },
                    "404": {
                        "description": "Coupon not found"
                    }
                }
            }
        },
        "/api/image": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get all images",
                "operationId": "1635eef9b114be63a3ac64e90831a3ae",
                "responses": {
                    "200": {
                        "description": "List of all images"
                    }
                }
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "upload a new images",
                "operationId": "1df06bb2db0319e20582f119b51b6eb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "img[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Images uploaded"
                    }
                }
            }
        },
        "/api/image/{id}": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get image by ID",
                "operationId": "4bba6f24ab0edb74e9fa3c4aa1b78996",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image found"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Delete an image by ID",
                "operationId": "ac560551c46b802a357cd0f01325efb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image delete successfully"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                }
            }
        },
...
    },
...
}